+++ zephyrapp/game/src/cowtools/cat_curves.h
@@ -7,17 +7,17 @@
 
 static inline float CAT_ease_in_sine(float t)
 {
-	return 1 - cos(0.5f * M_PI * t);
+	return 1 - cosf(0.5f * M_PI * t);
 }
 
 static inline float CAT_ease_out_sine(float t)
 {
-	return sin(0.5f * M_PI * t);
+	return sinf(0.5f * M_PI * t);
 }
 
 static inline float CAT_ease_inout_sine(float t)
 {
-	return -(cos(M_PI * t)-1) * 0.5f;
+	return -(cosf(M_PI * t)-1) * 0.5f;
 }
 
 static inline float CAT_ease_in_quad(float t)
@@ -45,7 +45,7 @@ static inline float CAT_ease_in_cubic(float t)
 
 static inline float CAT_ease_out_cubic(float t)
 {
-	return 1 - pow(1-t, 3);
+	return 1 - powf(1-t, 3);
 }
 
 static inline float CAT_ease_inout_cubic(float t)
@@ -53,7 +53,7 @@ static inline float CAT_ease_inout_cubic(float t)
 	return
 	t < 0.5f ?
 	4 * t*t*t :
-	1 - pow(-2*t+2, 3) * 0.5f;
+	1 - powf(-2*t+2, 3) * 0.5f;
 }
 
 static inline float CAT_ease_in_quart(float t)
@@ -63,14 +63,14 @@ static inline float CAT_ease_in_quart(float t)
 
 static inline float CAT_ease_out_quart(float t)
 {
-	return 1 - pow(1-t, 4);
+	return 1 - powf(1-t, 4);
 }
 
 static inline float CAT_ease_inout_quart(float t)
 {
 	return t < 0.5f ?
 	8 * t*t*t*t :
-	1 - pow(-2*t+2, 4) * 0.5f;
+	1 - powf(-2*t+2, 4) * 0.5f;
 }
 
 static inline float CAT_ease_in_quint(float t)
@@ -80,28 +80,28 @@ static inline float CAT_ease_in_quint(float t)
 
 static inline float CAT_ease_out_quint(float t)
 {
-	return 1 - pow(1-t, 5);
+	return 1 - powf(1-t, 5);
 }
 
 static inline float CAT_ease_inout_quint(float t)
 {
 	return t < 0.5f ?
 	16 * t*t*t*t*t :
-	1 - pow(-2*t+2, 5) * 0.5f;
+	1 - powf(-2*t+2, 5) * 0.5f;
 }
 
 static inline float CAT_ease_in_expo(float t)
 {
 	return
 	t == 0 ? 0 :
-	pow(2, 10*t-10);
+	powf(2, 10*t-10);
 }
 
 static inline float CAT_ease_out_expo(float t)
 {
 	return
 	t == 1 ? 1 :
-	1 - pow(2, -10*t);
+	1 - powf(2, -10*t);
 }
 
 static inline float CAT_ease_inout_expo(float t)
@@ -110,8 +110,8 @@ static inline float CAT_ease_inout_expo(float t)
 	t == 0 ? 0 :
 	t == 1 ? 1 :
 	t < 0.5f ?
-	pow(2, 20*t-10) * 0.5f :
-	(2 - pow(2, -20*t+10)) * 0.5f;
+	powf(2, 20*t-10) * 0.5f :
+	(2 - powf(2, -20*t+10)) * 0.5f;
 }
 
 static inline float CAT_ease_in_circ(float t)
@@ -143,7 +143,7 @@ static inline float CAT_ease_out_back(float t)
 {
 	const float c1 = 1.70158f;
 	const float c2 = c1 + 1;
-	return 1 + c2*pow(t-1, 3) + c1*pow(t-1, 2);
+	return 1 + c2*powf(t-1, 3) + c1*powf(t-1, 2);
 }
 
 static inline float CAT_ease_inout_back(float t)
@@ -162,7 +162,7 @@ static inline float CAT_ease_in_elastic(float t)
 	return
 	t == 0 ? 0 :
 	t == 1 ? 1 :
-	-pow(2, 10*t-10) * sin(c*(t*10-10.75f));
+	-powf(2, 10*t-10) * sinf(c*(t*10-10.75f));
 }
 
 static inline float CAT_ease_out_elastic(float t)
@@ -171,7 +171,7 @@ static inline float CAT_ease_out_elastic(float t)
 	return
 	t == 0 ? 0 :
 	t == 1 ? 1 :
-	pow(2, -10*t) * sin(c*(t*10-0.75)) + 1;
+	powf(2, -10*t) * sinf(c*(t*10-0.75)) + 1;
 }
 
 static inline float CAT_ease_inout_elastic(float t)
@@ -181,8 +181,8 @@ static inline float CAT_ease_inout_elastic(float t)
 	t == 0 ? 0 :
 	t == 1 ? 1 :
 	t < 0.5f ?
-	-pow(2, 20*t-10) * sin(c*(20*t-11.125)) * 0.5f :
-	pow(2, -20*t+10) * sin(c*(20*t-11.125)) * 0.5f + 1;
+	-powf(2, 20*t-10) * sinf(c*(20*t-11.125)) * 0.5f :
+	powf(2, -20*t+10) * sinf(c*(20*t-11.125)) * 0.5f + 1;
 }
